Create README.MD
HRMS is a web app that I started deveolping for a company that I used to work for, my plan was to help the company manage its employees as
easy as possible not only that but to stop using papers too! so the system supposed to connect all the organizations/departments together

One of the features I added to HRMS is tasks for employees. For example an employee could have a task in finishing some task with the govt
or an employee how who gives permission to leave work early, or approves vacation, etc. 

this web app is still under deveolpment....